+++ aux_codec.js
@@ -13,9 +13,9 @@ or implied.
 *
 * Repository: gve_devnet_webex_devices_executive_room_multi_aux_switching_macro
 * Macro file: aux_codec
-* Version: 1.0.19
-* Released: January 19, 2024
-* Latest RoomOS version tested: 11.11.1.9
+* Version: 1.0.20
+* Released: August 19, 2024
+* Latest RoomOS version tested: 11.19.1.7
 *
 * Macro Author:      	Gerardo Chaves
 *                    	Technical Solutions Architect